requirements.txt
# Core Web Framework
Django==5.2.4

# REST API Support
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0

# PostgreSQL Support
psycopg2==2.9.10

# Background Task Queue
celery==5.5.3
redis==6.2.0

# CORS Support
django-cors-headers==4.7.0

# Image Handling
pillow==11.3.0

# Environment Configuration
python-decouple==3.8
python-dotenv==1.1.1
